couldn't figure out whats wrong with my binary search and unsigned wrapping.
git-svn-id: https://subversion.ews.illinois.edu/svn/su15-cs225/lyan5/mp_qtree@6214 81a470a5-b7fd-438a-be91-513ba3ee436a

+++ include/quadtree.h
@@ -35,8 +35,8 @@ class quadtree
 	void rotate_clockwise();
 
 	void prune(unsigned tolerance);
-	unsigned pruned_size(unsigned tolerance)const;
-	unsigned ideal_prune(unsigned leaves)const;
+	uint64_t pruned_size(unsigned tolerance)const;
+	uint32_t ideal_prune(unsigned leaves)const;
 
   private:
     /**
